[ taha.darendeli ][ Implemented MVP Architecture ]
