+++ Clothes Store/Controllers/CatalogueScene/CatalogueViewController+CollectionView.swift
@@ -1,0 +1,42 @@
+//
+//  CatalogueViewController+CollectionView.swift
+//  Clothes Store
+//
+//  Created by Mentor on 1.06.2021.
+//  Copyright © 2021 RichieHope. All rights reserved.
+//
+
+import UIKit
+
+extension CatalogueViewController: UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout{
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return products.count
+    }
+
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: Strings.Identifiers.productCell.rawValue, for: indexPath) as! CatalogueViewCollectionViewCell
+
+        let product = products[indexPath.row]
+        cell.configureWithProduct(product: product)
+        return cell
+    }
+
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let itemSize = (collectionView.frame.width - (collectionView.contentInset.left + collectionView.contentInset.right + 10)) / 2
+        return CGSize(width: itemSize, height: 250)
+    }
+
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+
+        let product = products[indexPath.row]
+
+        let detailVC = self.storyboard?.instantiateViewController(withIdentifier: Strings.Identifiers.detailContainer.rawValue) as! DetailViewContainerViewController
+        let navigationVC = UINavigationController(rootViewController: detailVC)
+        detailVC.product = product
+        
+        Haptic.feedBack()
+        
+        self.present(navigationVC, animated: true, completion: nil)
+    }
+}
