+++ Clothes Store/Controllers/WishlistScene/WishlistViewController+TableView.swift
@@ -1,0 +1,56 @@
+//
+//  WishlistViewController+TableView.swift
+//  Clothes Store
+//
+//  Created by Mentor on 1.06.2021.
+//  Copyright © 2021 RichieHope. All rights reserved.
+//
+
+import UIKit
+
+extension WishlistViewController: UITableViewDelegate{
+
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return 125
+    }
+
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+
+        tableView.deselectRow(at: indexPath, animated: true)
+    }
+
+    func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
+
+        let deleteAction = UIContextualAction.init(style:.destructive, title:  Strings.Texts.remove.rawValue, handler: { [weak self] (action, view, completion) in
+            guard let self = self else { return }
+            
+            Haptic.feedBack()
+            
+            self.presenter.removeProductFromWishlist(product: self.products[indexPath.row])
+        })
+
+        deleteAction.backgroundColor = UIColor.primaryColour
+
+        let config = UISwipeActionsConfiguration(actions: [deleteAction])
+        config.performsFirstActionWithFullSwipe = true
+        return config
+
+    }
+}
+
+extension WishlistViewController: UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        products.count
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        if let cell = tableView.dequeueReusableCell(withIdentifier: Strings.Identifiers.wishlistCell.rawValue, for: indexPath) as? SavedViewTableViewCell {
+            cell.configureWithProduct(product: products[indexPath.row])
+            cell.delegate = self
+            
+            return cell
+        }
+        
+        return UITableViewCell()
+    }
+}
